structured book-recomendation system
